Move axis registry to own module
